move substitute to helpers.go
